refactor: reorganize api evaluation folders for semantics
