+++ src/front-end/typescript/lib/http/api/opportunity/sprint-with-us/team-questions/evaluations.ts
@@ -1,7 +1,7 @@
 import * as crud from "front-end/lib/http/crud";
 import * as Resource from "shared/lib/resources/evaluations/sprint-with-us/team-questions";
 import { Id } from "shared/lib/types";
-import { rawSWUTeamQuestionResponseEvaluationToSWUTeamQuestionResponseEvaluation } from "front-end/lib/http/api/question-evaluation/sprint-with-us";
+import { rawSWUTeamQuestionResponseEvaluationToSWUTeamQuestionResponseEvaluation } from "front-end/lib/http/api/evaluations/sprint-with-us/team-questions";
 
 const NAMESPACE = (opportunityId: Id) =>
   `opportunity/sprint-with-us/${opportunityId}/team-questions/evaluations`;
