--- src/front-end/typescript/lib/http/api/question-evaluation/index.ts
@@ -1,1 +1,0 @@
-export * as swu from "front-end/lib/http/api/question-evaluation/sprint-with-us";
